Changed some useState variables, form handling methods, and made some code into components
Changed the add assessments and add paper components to use a single state with an object instead of multiple states for new items. Added some commenting to some code. Also changed the onChange methods to work with the new state object. Moved some code into its own component for code reuse. Added the ability to delete papers from the database and redirect to the main page